The USA dropped Freddy Fazbear in Afghanistan.

+++ src/day7/mod.rs
@@ -9,7 +9,7 @@ pub fn part1() -> i32 {
     let median = v[v.len() / 2];
 
     let mut consume = 0;
-    v.iter().for_each(|i| consume += (*i - median as i32).abs());
+    v.iter().for_each(|i| consume += (*i - median).abs());
     consume
 }
 
@@ -21,7 +21,7 @@ pub fn part2() -> i32 {
     let avg = v.iter().sum::<i32>() / v.len() as i32;
     let mut consume = 0;
     v.iter().for_each(|v| {
-        let p = (*v - avg as i32).abs();
+        let p = (*v - avg).abs();
         for p in 0..p + 1 {
             consume += p;
         }
